FileAPI/blob/blob-in-worker.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Create Blob in Worker</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/FileAPI/#blob">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

async_test(function() {
  var data = "TEST";
  var worker = new Worker("../support/blob-in-worker.js");

  worker.onmessage = this.step_func(function(event) {
    assert_equals(event.data, data);
    this.done();
  });

  worker.onerror = this.step_func(function(event) {
    assert_unreached(event.message);
  });
  worker.postMessage(data);

})

</script>
